Move libs.os_helpers to devtools.utils and fix imports
