--- libs/os_helpers/pyproject.toml
@@ -1,12 +1,0 @@
-[tool.poetry]
-name = "os_helpers"
-version = "0.1.0"
-description = "Helper classes for low level OS programming."
-authors = ["Felix Trautwein <mail@felixtrautwein.de>"]
-
-[tool.poetry.dependencies]
-python="^3.10"
-
-[build-system]
-requires = ["poetry>=0.12"]
-build-backend = "poetry.core.masonry.api"
