+++ libs/readme.md
@@ -1,0 +1,1 @@
+Directory where libs are created (see `dev create --help`).
