+++ tools/devtools/src/devtools/targets/tool/tool.py
@@ -2,10 +2,9 @@ import logging
 from pathlib import Path
 
 from devtools.config import REPO_CONFIG
-from devtools.targets import util
+from devtools.targets import target
 from devtools.targets.target import Target
-
-from wrappers.pytest_wrapper import pytest
+from devtools.wrappers import pytest
 
 logger = logging.getLogger()
 
@@ -13,9 +12,7 @@ logger = logging.getLogger()
 class Tool(Target):
     """Python command line application."""
 
-    def __init__(
-        self, target_name: str, base_target_dir: Path, base_build_dir: Path
-    ) -> None:
+    def __init__(self, target_name: str, base_target_dir: Path, base_build_dir: Path) -> None:
         Target.__init__(self, target_name, base_target_dir, base_build_dir)
 
     @classmethod
@@ -28,20 +25,15 @@ class Tool(Target):
         (target_src_dir / "__init__.py").touch()
         target_tests_dir = target_dir / "tests"
         target_tests_dir.mkdir(parents=True)
-        util.extend_pythonpath(target_src_dir)
-        logger.info(
-            "Your target %s was set up, please register it in registered_targets.py.",
-            target_name,
-        )
+        target.extend_pythonpath(target_src_dir)
+        logger.info("Your target %s was set up, please register it in registered_targets.py.", target_name)
 
     def build(self, clean: bool) -> None:
         raise NotImplementedError
 
     def test(self, coverage: bool) -> bool:
         coverage_dir = REPO_CONFIG.metrics_dir / "pytest-coverage" / self.target_name
-        return pytest.test_directory(
-            self.target_tests_dir, out_dir=coverage_dir, coverage=coverage, vis=False
-        )
+        return pytest.test_directory(self.target_tests_dir, out_dir=coverage_dir, coverage=coverage, vis=False)
 
     def deploy(self) -> None:
         raise NotImplementedError
