--- tools/devtools/src/devtools/util.py
@@ -1,5 +1,0 @@
-"""Simple utility functions for devtools."""
-
-
-def stringify(commands: list[str]) -> str:
-    return " ".join(commands)
