+++ tools/devtools/tests/test_utils.py
@@ -3,8 +3,7 @@ import os
 from typing import NoReturn
 
 import pytest
-
-from os_helpers import os_helpers
+from devtools import utils
 
 
 def callable_raising_exception() -> NoReturn:
@@ -14,18 +13,18 @@ def callable_raising_exception() -> NoReturn:
 def test_wait_for_process() -> None:
     """Tests exception for process with non-zero exit code."""
     with pytest.raises(ValueError):
-        with os_helpers.wait_for_callable(callable_raising_exception):
+        with utils.wait_for_callable(callable_raising_exception):
             pass
 
 
 def test_temp_env() -> None:
     """Test environment state when using temp_env context manager."""
     old_env = dict(os.environ)
-    with os_helpers.temp_env(test="test_variable"):
+    with utils.temp_env(test="test_variable"):
         assert os.environ.get("test") == "test_variable"
         os.environ["test2"] = "test_variable"
         assert os.environ.get("test2") == "test_variable"
     assert old_env == dict(os.environ)
-    with os_helpers.temp_env(all_caps=True, test="test_variable"):
+    with utils.temp_env(all_caps=True, test="test_variable"):
         assert os.environ.get("TEST") == "test_variable"
         assert "test" not in os.environ
